Move FDA & HistogramMatching to domain_adaptation.py
